Introduce new process activate/deactivate scheme.
To help manage the transfer between thread private heaps and process
heaps, new "activate" and "deactivate" functions are added
and incorporated.

This fixes a segfaulting bug that was caused by using an invalid
thread private heap. The heap was invalid because it wasn't restored
after a OOM condition was hit. Things should be better now.