Fix out-of-control memory use caused by trusting the QUOTA_HIT tag too much.
Under just the right conditions, a producer will consume all the input for a given
round without ever needing to request a refresh buffer (via nile_Process_append_output).
This means that the QUOTA_HIT tag will never be set. Before this fix, we relied solely
on that tag to determine when to wake-up or block on the consumer. Now we
check on the consumer at the end of each round, regardless of if the tag is set.