Fix Reverse process (previous commit broke it).
The previous version of Reverse relied on nile_Process_run
to ignore the consumer's input queue as long as the QUOTA_HIT tag
was not set on a buffer. This is because Reverse used the
consumer's input as intermediate storage for the reversed
stream. Because nile_Process_run now looks a the consumer's
input queue to determine if the consumer is ready to run,
the new version of Reverse must keep the intermediate storage
elsewhere (in a process variable).