Redo out-of-memory fix.
The previous approach would check consumer quota even if the tag was not set.
This is a bit too "nosy"; we need to allow processes to have "unmanaged" consumers.
This new approach trusts the tag as before, but is careful to set the tag
at the start of nile_Process_run.