Fix out-of-control pipeline creation.
With enough threads, and a pipeline that doesn't have enough
parallelism, we can end up with many threads that are far ahead
the other threads working on the oldest pipeline. This can result
in out-of-control pipeline creation because we allowed the
liaison to continue to bring in more pipelines when it couldn't
find work to do. Now we make the worker doze if work is not
found. The liaison can only resume if the variable truly drops
below the value.