New OS thread sleeping approach.
Designed to reduce unnecessary semaphore signal/waiting.
This is probably the cause of the excessive fan noise
when running a nile program. The fan is on because the
various threads (and thus cores) are busy working
through the (sometimes thousands of) semaphore signals
that have built up.

We no longer signal the semaphore everytime there is work to
be done and a thread is blocked on the semaphore. If we have
already signaled the semaphore enough to wake up all blocked
threads, we no longer signal.

And, instead of a quiescent semaphore, we just monitor
the number of sleeping threads, dozing as needed to
reduce polling.