The JS AST is emitting the string "nodePrototype" all over the place resulting in a wonky AST.
e.g. Using the hard-coded gezira code inside nile-compiler.html, after
hitting compile inspect the right hand textarea, and see:

opdef {
   sig=opsig {
       name=nodePrototype
       fixity=in
       param=[
           nodePrototype:"nodePrototype"           nodePrototype:"nodePrototype"       ]
       type="nodePrototype"       CharacterRange=82-113
   }
   body=<null>
   CharacterRange=82-113
}

Looking at the character range in the defs for this reveals:

> defs.substring(82,113)
"a:Boolean ∨ b:Boolean : Boolean"

The problem stems from JS functions having a name property populated on them, e.g.:

> foobar = function() {}
< function () {}
> foobar.name
< "foobar"

So any of the AST nodes with a name property end up displaying "nodePrototype".

A quick/easy fix is to override the setter/getters for the name property
on nodePrototype, and re-route it to a __name__ property.

With the fix, the AST looks proper again:

opdef {
   sig=opsig {
       name=∨
       fixity=in
       param=[
           a:"Boolean"           b:"Boolean"       ]
       type="Boolean"       CharacterRange=82-113
   }
   body=<null>
   CharacterRange=82-113
}